Enhance monolithic build options and improve logging functionality
- Added new CMake options for Tracy profiler instrumentation and link-time optimization for monolithic builds, allowing for better performance tuning.
- Updated function signatures in `vmMain` across various modules to use `intptr_t` for argument types, ensuring compatibility with 64-bit architectures.
- Improved logging mechanisms by introducing structured logging capabilities and enhancing error handling in various subsystems, including sound and networking.
- Added a new ENet loopback test to validate network functionality when using the ENet library.

These changes collectively enhance the build process for monolithic configurations and improve the overall logging and network testing capabilities within the game.

+++ mymod/gamesrc/cgame/cg_main.c
@@ -53,7 +53,7 @@ This must be the very first function compiled into the .q3vm file
 extern intptr_t (QDECL *syscall)( intptr_t arg, ... );
 #endif
 #ifdef COMBINED_MONOLITH
-static intptr_t QDECL vmMain(int command, int arg0, int arg1, int arg2, [[maybe_unused]] int arg3, [[maybe_unused]] int arg4, [[maybe_unused]] int arg5, [[maybe_unused]] int arg6, [[maybe_unused]] int arg7, [[maybe_unused]] int arg8, [[maybe_unused]] int arg9, [[maybe_unused]] int arg10, [[maybe_unused]] int arg11) {
+static intptr_t QDECL vmMain(int command, intptr_t arg0, intptr_t arg1, intptr_t arg2, [[maybe_unused]] intptr_t arg3, [[maybe_unused]] intptr_t arg4, [[maybe_unused]] intptr_t arg5, [[maybe_unused]] intptr_t arg6, [[maybe_unused]] intptr_t arg7, [[maybe_unused]] intptr_t arg8, [[maybe_unused]] intptr_t arg9, [[maybe_unused]] intptr_t arg10, [[maybe_unused]] intptr_t arg11) {
 #else
 Q_EXPORT intptr_t vmMain(int command, int arg0, int arg1, int arg2, [[maybe_unused]] int arg3, [[maybe_unused]] int arg4, [[maybe_unused]] int arg5, [[maybe_unused]] int arg6, [[maybe_unused]] int arg7, [[maybe_unused]] int arg8, [[maybe_unused]] int arg9, [[maybe_unused]] int arg10, [[maybe_unused]] int arg11) {
 #endif
